+++ sigma/plugins/administration/plugin.yml
@@ -177,15 +177,6 @@ commands:
     description:
       Toggles if a server by the given ID is or is not an Aurora Project partner.
 
-  - name:    collectchain
-    global:  true
-    enabled: true
-    admin:   true
-    sfw:     true
-    usage:   "{pfx:s}{cmd:s} @person"
-    description:
-      Collects messages sent by the mentioned user and saves it as a chain.
-
 events:
   - name:    pmredirect
     type:    message
